ref: rename mod gen -> util
We can now stuff more util stuff in this file :)
